+++ src/di/technicianModule.ts
@@ -11,8 +11,8 @@ import { EmailService } from "../utils/email";
 import { EmailTemplateService } from "../utils/emailTemplates";
 import { OTPService } from "../utils/otp";
 import { RedisService } from "../utils/redis";
-import { ItechnicianAuthService } from "../interfaces/Iservices/ItechnicianAuthService";
-import { ItechnicianService } from "../interfaces/Iservices/ItechnicianService";
+import { ItechnicianAuthService } from "../interfaces/Iservices/ItechnicianService/ItechnicianAuthService";
+import { ItechnicianService } from "../interfaces/Iservices/ItechnicianService/ItechnicianService";
 import { ItechnicianRepository } from "../interfaces/Irepositories/ItechnicianRepository";
 import { ItempTechnicianRepository } from "../interfaces/Irepositories/ItempTechnicianRepository";
 import { IjobDesignationRepository } from "../interfaces/Irepositories/IjobDesignationRepository";
