--- src/interfaces/Iservices/IadminService.ts
@@ -1,6 +1,0 @@
-import { loginResponseDTO } from "../DTO/IServices/adminService.dto";
-import { loginDataDTO } from "../DTO/IServices/adminService.dto";
-
-export interface IadminService{
-    adminLogin(data:loginDataDTO):Promise<loginResponseDTO>
-}
