+++ src/interfaces/Iservices/ItechnicianService/ItechnicianService.ts
@@ -1,4 +1,4 @@
-import { getJobDesignationsResponse } from "../DTO/IServices/technicianService.dto";
+import { getJobDesignationsResponse } from "../../DTO/IServices/technicianService.dto";
 
 export interface ItechnicianService {
     getJobDesignations(): Promise<getJobDesignationsResponse>;
