+++ src/services/adminService/jobDesignationService.ts
@@ -1,5 +1,5 @@
 import { injectable, inject } from "tsyringe";
-import { IjobDesignationService } from "../../interfaces/Iservices/IjobDesignationService";
+import { IjobDesignationService } from "../../interfaces/Iservices/IadminService/IjobDesignationService";
 import { IjobDesignationRepository } from "../../interfaces/Irepositories/IjobDesignationRepository";
 import { AddDesignationResponseDTO } from "../../interfaces/DTO/IServices/jobDesignationService.dto";
 import { HTTP_STATUS } from "../../utils/httpStatus";
